Pass setTaskAnswer callback to Question
