+++ ultrack-prototype/frontend/components/Question.tsx
@@ -1,26 +1,14 @@
 import { Button } from "@czi-sds/components";
 import { Box, Typography } from "@mui/material";
 import { Answer, Task } from "../task";
-import { Dispatch, SetStateAction } from "react";
 
 export type QuestionProps = {
-  taskIndex: number;
-  tasks: Task[];
-  setTasks: Dispatch<SetStateAction<Task[]>>;
-  setTaskIndex: Dispatch<SetStateAction<number>>;
+  task: Task;
+  setTaskAnswer: (answer: Answer) => void;
 };
 
 export default function Question(props: QuestionProps) {
-  const { taskIndex, tasks, setTasks, setTaskIndex } = props;
-  const task = tasks[taskIndex];
-
-  const setAnswer = (answer: Answer) => {
-    const updatedTask = structuredClone(task);
-    updatedTask.answer = answer;
-    const updatedTasks = Array(...tasks);
-    updatedTasks[taskIndex] = updatedTask;
-    setTasks(updatedTasks);
-  };
+  const { task, setTaskAnswer } = props;
 
   return (
     <Box
@@ -47,10 +35,7 @@ export default function Question(props: QuestionProps) {
             sdsStyle="square"
             onClick={() => {
               const key = answer as keyof typeof Answer;
-              setAnswer(Answer[key]);
-              setTaskIndex((prevIndex) =>
-                prevIndex < tasks.length - 1 ? prevIndex + 1 : prevIndex
-              );
+              setTaskAnswer(Answer[key]);
             }}
           >
             {label}
